Use *-panel icons for the statusicon if requested
This adds proper support for the audio-volume-*-panel icon variants. If
requested in the config, we now try to use audio-volume-*-panel for the
statusicon if the default GTK icon theme is used. Otherwise we fall back
on the regular audio-volume-* variants. Note that the notification icons
always use the system's GTK icon theme again now. We mistakenly changed
this to use the *-panel versions as well at some point.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>